Moves offset data function to yaml utilities
